Add a CRAN Status badge

This function adds a CRAN Status badge to the README.Rmd. If the package is not hosted on the CRAN the badge will indicate not published on the CRAN.

Make sure that 1) a README.Rmd file exists at the project root and 2) it contains a block starting with the line <!-- badges: start --> and ending with the line <!-- badges: end -->.

Don't forget to re-render the README.md.

add_cran_badge(quiet = FALSE)

Arguments

quiet

A logical value. If TRUE messages are deleted. Default is FALSE.

Value

A badge as a markdown expression.

See also

Examples

if (FALSE) {
add_cran_badge()
}