Add a Dependencies badge

This function adds or updates the Dependencies badge to the README.Rmd. The first number corresponds to the direct dependencies and the second to the recursive dependencies.

Note: this function can work with packages not published on the CRAN and is based on the function gtools::getDependencies(). See also the function get_all_dependencies().

Make sure that 1) a README.Rmd file exists at the project root and 2) it contains a block starting with the line <!-- badges: start --> and ending with the line <!-- badges: end -->.

Don't forget to re-render the README.md.

Usage

add_dependencies_badge(quiet = FALSE)

Arguments

quiet

A logical value. If TRUE messages are deleted. Default is FALSE.

Value

A badge as a markdown expression.

See also

Examples

if (FALSE) {
add_dependencies_badge()
}