Add a R CMD Check badge

This function adds a R CMD Check badge to the README.Rmd. This function must be run after add_github_actions_check() which will setup GitHub Actions to check and test the package.

Make sure that 1) a README.Rmd file exists at the project root and 2) it contains a block starting with the line <!-- badges: start --> and ending with the line <!-- badges: end -->.

Don't forget to re-render the README.md.

Usage

add_github_actions_check_badge(organisation = NULL, quiet = FALSE)

Arguments

organisation

A character of length 1. The name of the GitHub organisation to host the package. If NULL (default) the GitHub account will be used.

quiet

A logical value. If TRUE messages are deleted. Default is FALSE.

Value

A badge as a markdown expression.

See also

Examples

if (FALSE) { # \dontrun{
add_github_actions_check_badge()
} # }