Create a README file

This function creates a README.Rmd file at the root of the project based on a template. Once edited user needs to knit it into a README.md (or use the function refresh()).

add_readme_rmd(
  type = "package",
  given = NULL,
  family = NULL,
  organisation = NULL,
  open = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)

Arguments

type

A character of length 1. If package (default) a GitHub README.Rmd specific to an R package will be created. If compendium a GitHub README.Rmd specific to a research compendium will be created.

given

A character of length 1. The given name of the project maintainer.

family

A character of length 1. The family name of the project maintainer.

organisation

A character of length 1. The name of the GitHub organisation to host the package. If NULL (default) the GitHub account will be used. This argument is used to set the URL of the package (hosted on GitHub).

open

A logical value. If TRUE (default) the file is opened in the editor.

overwrite

A logical value. If this file is already present and overwrite = TRUE, it will be erased and replaced. Default is FALSE.

quiet

A logical value. If TRUE messages are deleted. Default is FALSE.

Value

No return value.

See also

Examples

if (FALSE) {
add_readme_rmd(type = "package")
}