Create an R package structure

This function creates a new R package structure according to the CRAN policies. Essential features of an R package are created (DESCRIPTION and NAMESPACE files, and R/ and man/ folders). The project is also versioning with git and a generic R .gitignore is added.

IMPORTANT - Before using this function user needs to create a new folder (or a new project if using RStudio) and run this function inside this folder (by using setwd() or by opening the Rproj in a new Rstudio session). The name of the package will be the same as the name of this folder and some rules must be respected: https://r-pkgs.org/workflows101.html#naming.

Some fields of the DESCRIPTION file (maintainer information, package name, license, URLs, and roxygen2 version) are automatically filled but others (title and description) need to be edited manually.

Additional features are also created: a CITATION file, a README.Rmd, and a vignettes/ folder (optional). See below the section Package Content to see the full structure of the R package.

A GitHub repository can also be created (default) following the "GitHub last" workflow (https://happygitwithr.com/existing-github-last.html). Configuration files for GitHub Actions to automatically check the package and deploy the website pkgdown::pkgdown() will be added in the .github/ folder. See below the section Creating a GITHUB Repository.

new_package(
  license = "GPL (>= 2)",
  status = "concept",
  lifecycle = "experimental",
  create_repo = TRUE,
  private = FALSE,
  gh_check = TRUE,
  website = TRUE,
  given = NULL,
  family = NULL,
  email = NULL,
  orcid = NULL,
  github = NULL,
  organisation = NULL,
  overwrite = FALSE,
  quiet = FALSE
)

Arguments

license
a character vector of length 1

The license to be used for this package. Run get_licenses() to choose an appropriate one. Default is license = 'GPL (>= 2)'

The license can be changed later by using add_license() (and add_license_badge() or refresh() to update the corresponding badge).

status
a character vector of length 1

The status of the project according to the standard defined by the https://www.repostatus.org/ project. One among 'concept' (default), 'wip', 'suspended', 'abandoned', 'active', 'inactive', or 'unsupported'. See add_repostatus_badge() for further information.

This argument is used to add a badge to the README.Rmd to help visitors to better understand your project. If you don't want this badge use status = NULL.

This status can be added/changed later by using add_repostatus_badge().

lifecycle
a character vector of length 1

The lifecycle stage of the project according to the standard defined at https://lifecycle.r-lib.org/articles/stages.html. One among 'experimental' (default), 'stable', 'deprecated', or 'superseded'. See add_lifecycle_badge() for further information.

This argument is used to add a badge to the README.Rmd to help visitors to better understand your project. If you don't want this badge use lifecycle = NULL.

This stage can be added/changed later by using add_lifecycle_badge().

create_repo
a logical value

If TRUE (default) creates a repository (public if private = FALSE or private if private = TRUE) on GitHub. See below the section Creating a GITHUB Repository.

private
a logical value

If TRUE creates a private repository on user GitHub account (or organisation). Default is private = FALSE.

gh_check
a logical value

If TRUE (default) configures GitHub Actions to automatically check and test the package after each push. This will run R CMD check on the three major operating systems (Ubuntu, macOS, and Windows) on the latest release of R. See add_github_actions_check() for further information.

If create_repo = FALSE this argument is ignored.

website
a logical value

If TRUE (default) configures GitHub Actions to automatically build and deploy the package website (using pkgdown::pkgdown()) after each push. A gh-pages branch will be created using usethis::use_github_pages() and the GitHub repository will be automatically configured.

If create_repo = FALSE this argument is ignored.

given
a character vector of length 1

The given name of the maintainer of the package. If NULL (default) the function will try to get this value by reading the .Rprofile file.

For further information see set_credentials() and below the section Managing Credentials.

family
a character vector of length 1

The family name of the maintainer of the package. If NULL (default) the function will try to get this value by reading the .Rprofile file.

For further information see set_credentials() and below the section Managing Credentials.

email
a character vector of length 1

The email address of the maintainer of the package. If NULL (default) the function will try to get this value by reading the .Rprofile file.

For further information see set_credentials() and below the section Managing Credentials.

orcid
a character vector of length 1

The ORCID of the maintainer of the package. If NULL (default) the function will try to get this value by reading the .Rprofile file.

For further information see set_credentials() and below the section Managing Credentials.

github
a character vector of length 1

The GitHub pseudo of the maintainer of the package. It will be used to create the repository on GitHub unless !is.null(organisation) (the repository will be created on the GitHub organisation specified by the argument organisation).

If NULL (default) the function will try to get this value by reading the .Rprofile file (unless !is.null(organisation)).

organisation
a character vector of length 1

The GitHub organisation to host the repository. If defined it will overwrite the argument github.

Default is organisation = NULL (the GitHub pseudo will be used).

overwrite
a logical value

If TRUE files written from templates and modified by user are erased. Default is overwrite = FALSE. Be careful while using this argument.

quiet
a logical value

If TRUE messages are deleted. Default is FALSE.

Details

Package Content

With default settings the package pkg will contain:

.
│
├── .git/                     # GIT tracking folder
├── .github/                  # (optional) GitHub Actions settings
│   └── workflows/
│       ├── pkgdown.yaml      # Configuration file to build & deploy website
│       └── R-CMD-check.yaml  # Configuration file to check & test package
├── .Rbuildignore             # Automatically generated
├── .gitignore                # Specific to R packages
├── pkg.Rproj                 # (optional) Created by user 
├── _pkgdown.yaml             # (optional) User website settings
│
├── R/                        # R functions
│   └── pkg-package.R         # Dummy R file for package-level documentation
│
├── man/                      # R functions helps (automatically updated)
│   ├── figures/              # Figures for the README 
│       └── hexsticker.png    # Template R package HexSticker
│   └── pkg-package.Rd        # Package-level documentation
│
├── inst/
│   └── CITATION              # BiBTeX entry to cite the R package       [*]
│
├── vignettes/
│   └── pkg_vignette.Rmd      # (optional) Package tutorial              [*]
│
├── DESCRIPTION               # Project metadata                         [*]
├── NAMESPACE                 # Automatically generated
│
├── LICENSE                   # (optional) If License = MIT
├── LICENSE.md                # Content of the chosen license
│
├── README.md                 # GitHub README (automatically generated)
└── README.Rmd                # GitHub README (to knit)                  [*]

[*] These files are automatically edited but user needs to add manually 
    some information.

Recommended Workflow

The purpose of the package rcompendium is to make easier the creation of R package/research compendium so that user can focus on the code/analysis instead of wasting time organizing files.

The recommended workflow is as follow:

Create an empty RStudio project;

Store your credentials with set_credentials() (if not already done);

Run new_package() to create the package structure (and the GitHub repository);

Edit some metadata in DESCRIPTION, CITATION, and README.Rmd;

Implement, document & test functions (the fun part);

Update the project (update .Rd files, NAMESPACE, external dependencies in DESCRIPTION, re-knit README.Rmd, and check package integrity) with refresh().

Managing Credentials

You can use the arguments given, family, email, orcid, and github directly with the function new_package() (and others). But if you create a lot a projects (packages and/or compendiums) it can be frustrating in the long run.

An alternative is to use ONCE AND FOR ALL the function set_credentials() to permanently store this information in the .Rprofile file. If these arguments are set to NULL (default) each function of the package rcompendium will search in this .Rprofile file. It will save your time (it's the purpose of this package).

Even if you have stored your information in the .Rprofile file you will always be able to modify them on-the-fly (i.e. by using arguments of the new_package()) or permanently by re-running set_credentials().

Creating a GITHUB Repository

First run gh::gh_whoami() to see if your git and associated credentials are correctly configured. If so you should see something like:

{
"name": "Given Family",
"login": "pseudo",
"html_url": "https://github.com/pseudo",
...
}

Otherwise see gert::git_config() to debug.

To create the GitHub repository directly from R, the package rcompendium uses the function usethis::use_github(), an interface to the GitHub REST API. The interaction with this API required an authentication method: a GITHUB PAT (Personal Access Token).

If you don't have a GITHUB PAT locally stored, you must:

Obtain a new one from your GitHub account. Make sure to select at least the first two scopes (private repository and workflow)

Store it in the .Renviron file by using usethis::edit_r_environ() and adding the following line: GITHUB_PAT='99z9...z9'

See usethis::gh_token_help() for more information about getting and configuring a GITHUB PAT.

If everything is well configured you should see something like:

{
"name": "Given Family",
"login": "pseudo",
"html_url": "https://github.com/pseudo",
"scopes": "delete_repo, repo, workflow",
"token": "99z9...z9"
}

And you will be able to create a GitHub repository directly from R!

See also

Examples

if (FALSE) {
library(rcompendium)

## Define **ONCE FOR ALL** your credentials ----
set_credentials("Given", "Family", "email.address@domain.com", 
                github = "pseudo", orcid = "0000-0000-0000-0000")

## Create an R package ----
new_package()

## Start developing functions ----
## ...

## Update package (documentation, dependencies, README, check) ----
refresh()
}