Refresh a research compendium/package

This function refreshes a research compendium/package. It will:

Update Rd files and NAMESPACE, using devtools::document();

Update external packages (DESCRIPTION);

Update badges in README.Rmd (if already present);

Re-knitr the README.Rmd, using rmarkdown::render();

Check package integrity with devtools::check().

refresh(quiet = FALSE, check = TRUE, import = NULL)

Arguments

quiet
a logical value. If TRUE (default) no message are printed.

check
a logical value. If TRUE (default) package integrity is checked using devtools::check().

import
(optional) the name of the folder to recursively detect dependencies to be added in the Imports field of DESCRIPTION file. Default is import = NULL (but R/, NAMESPACE, and @examples are still inspected).

See also

Examples

if (FALSE) {
rcompendium::refresh()
}