Store credentials to the .RProfile

This function adds user credentials in the .Rprofile file. Accepted credentials are listed below. This function is useful if user creates a lot of packages and/or research compendiums.

If the .Rprofile file does not exist this function will create it. Note that credentials are added at the end of the file: if you run this function two times, credentials will be added twice. User can run the command usethis::edit_r_profile() to open the .Rprofile and clean its content. Be careful.

set_credentials(
  given = NULL,
  family = NULL,
  short = NULL,
  email = NULL,
  orcid = NULL,
  github = NULL,
  license = NULL
)

Arguments

given
(optional) user given name (i.e. "Nicolas")

family
(optional) user family name (i.e. "Casajus")

short
(optional) user short name (e.g. "Casajus N.")

email
(optional) user email address (i.e. "nicolas.casajus@gmail.com")

orcid
(optional) user ORCID (i.e. "0000-0002-5537-5294")

github
(optional) user GitHub pseudo/organization (i.e. "ahasverus")

license
(optional) user preferred license (i.e. "GPL (>= 2)")

See also

Examples

if (FALSE) {
set_credentials(given = "Nicolas", family = "Casajus", short = "Casajus N.",
                email = "nicolas.casajus@gmail.com", github = "ahasverus", 
                orcid = "0000-0002-5537-5294", license = "GPL (>= 2)")
}