Store credentials to the .RProfile

This function stores user credentials in the .Rprofile file. Accepted credentials are listed below. This function is useful if user creates a lot of packages and/or research compendiums.

If the .Rprofile file does not exist this function will create it. Users need to paste the content of the clipboard to this file.

set_credentials(
  given = NULL,
  family = NULL,
  email = NULL,
  orcid = NULL,
  protocol = NULL
)

Arguments

given
a character of length 1

The given name of the package maintainer.

family
a character of length 1

The family name of the package maintainer.

email
a character of length 1

The email address of the package maintainer.

orcid
a character of length 1

The ORCID of the package maintainer.

protocol
a character of length 1

The GIT protocol used to communicate with the GitHub remote. One of 'https' or 'ssh'. If you don't know, keep the default value (i.e. NULL) and the protocol will be 'https'.

Value

None

See also

Examples

if (FALSE) {
library(rcompendium)

## Define **ONCE FOR ALL** your credentials ----

set_credentials("John", "Doe", "john.doe@domain.com", 
                orcid = "9999-9999-9999-9999", protocol = "https")
}